+++ pages/core-units.tsx
@@ -1,5 +1,4 @@
-import { CuTable } from '@ses/containers/CUTable/CuTable';
-
+import { CuTable } from '@/views/CUTable/CuTable';
 import type { NextPage } from 'next';
 
 const CuTablePage: NextPage = () => <CuTable />;
