+++ src/views/CUTable/components/CustomTable/CardItemCoreUnitMobile/CardItemCoreUnitMobile.tsx
