+++ src/views/CUTable/components/CustomTable/ItemCoreUnit/ItemCoreUnit.tsx
